+++ tests/src/test/java/eu/fraho/blm2/st/BankTests.java
@@ -29,7 +29,7 @@ public class BankTests extends AbstractTest {
         driver.findElement(By.id("link_bank")).click();
 
         // check current balance
-        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50.000,00 €");
+        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50,000.00 €");
 
         // deposit 5k
         setValue(By.id("betrag"), "5000");
@@ -37,10 +37,10 @@ public class BankTests extends AbstractTest {
         assertElementPresent(By.id("meldung_207"));
 
         // check new balance
-        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 55.000,00 €");
+        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 55,000.00 €");
         Assertions.assertEquals("95000", driver.findElement(By.id("betrag")).getAttribute("value"));
-        assertText(By.id("stat_money"), "95.000,00 €");
-        assertText(By.id("stat_bank"), "55.000,00 €");
+        assertText(By.id("stat_money"), "95,000.00 €");
+        assertText(By.id("stat_bank"), "55,000.00 €");
     }
 
     @Test
@@ -52,10 +52,10 @@ public class BankTests extends AbstractTest {
         driver.findElement(By.id("do_transaction")).click();
         assertElementPresent(By.id("meldung_207"));
 
-        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 100.000,00 €");
+        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 100,000.00 €");
         Assertions.assertEquals("50000", driver.findElement(By.id("betrag")).getAttribute("value"));
-        assertText(By.id("stat_money"), "50.000,00 €");
-        assertText(By.id("stat_bank"), "100.000,00 €");
+        assertText(By.id("stat_money"), "50,000.00 €");
+        assertText(By.id("stat_bank"), "100,000.00 €");
     }
 
     @Test
@@ -66,10 +66,10 @@ public class BankTests extends AbstractTest {
         setValue(By.id("betrag"), "-100");
         driver.findElement(By.id("do_transaction")).click();
 
-        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50.000,00 €");
+        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50,000.00 €");
         Assertions.assertEquals("-100", driver.findElement(By.id("betrag")).getAttribute("value"));
-        assertText(By.id("stat_money"), "100.000,00 €");
-        assertText(By.id("stat_bank"), "50.000,00 €");
+        assertText(By.id("stat_money"), "100,000.00 €");
+        assertText(By.id("stat_bank"), "50,000.00 €");
     }
 
     @Test
@@ -81,10 +81,10 @@ public class BankTests extends AbstractTest {
         driver.findElement(By.id("do_transaction")).click();
         assertElementPresent(By.id("meldung_110"));
 
-        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50.000,00 €");
+        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50,000.00 €");
         Assertions.assertEquals("50000.01", driver.findElement(By.id("betrag")).getAttribute("value"));
-        assertText(By.id("stat_money"), "100.000,00 €");
-        assertText(By.id("stat_bank"), "50.000,00 €");
+        assertText(By.id("stat_money"), "100,000.00 €");
+        assertText(By.id("stat_bank"), "50,000.00 €");
     }
 
     @Test
@@ -98,9 +98,9 @@ public class BankTests extends AbstractTest {
         assertElementPresent(By.id("meldung_110"));
 
         // check new balance
-        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50.000,00 €");
-        assertText(By.id("stat_money"), "100.000,00 €");
-        assertText(By.id("stat_bank"), "50.000,00 €");
+        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50,000.00 €");
+        assertText(By.id("stat_money"), "100,000.00 €");
+        assertText(By.id("stat_bank"), "50,000.00 €");
     }
 
     @Test
@@ -115,9 +115,9 @@ public class BankTests extends AbstractTest {
         assertElementPresent(By.id("meldung_207"));
 
         // check new balance
-        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 45.000,00 €");
-        assertText(By.id("stat_money"), "105.000,00 €");
-        assertText(By.id("stat_bank"), "45.000,00 €");
+        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 45,000.00 €");
+        assertText(By.id("stat_money"), "105,000.00 €");
+        assertText(By.id("stat_bank"), "45,000.00 €");
     }
 
     @Test
@@ -132,9 +132,9 @@ public class BankTests extends AbstractTest {
         assertElementPresent(By.id("meldung_207"));
 
         // check new balance
-        assertText(By.id("cur_bank_account"), "Ihr Kontostand: -15.000,00 €");
-        assertText(By.id("stat_money"), "165.000,00 €");
-        assertText(By.id("stat_bank"), "-15.000,00 €");
+        assertText(By.id("cur_bank_account"), "Ihr Kontostand: -15,000.00 €");
+        assertText(By.id("stat_money"), "165,000.00 €");
+        assertText(By.id("stat_bank"), "-15,000.00 €");
     }
 
     @Test
@@ -149,10 +149,10 @@ public class BankTests extends AbstractTest {
         assertElementPresent(By.id("meldung_109"));
 
         // check new balance
-        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50.000,00 €");
+        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50,000.00 €");
         Assertions.assertEquals("65000.01", driver.findElement(By.id("betrag")).getAttribute("value"));
-        assertText(By.id("stat_money"), "100.000,00 €");
-        assertText(By.id("stat_bank"), "50.000,00 €");
+        assertText(By.id("stat_money"), "100,000.00 €");
+        assertText(By.id("stat_bank"), "50,000.00 €");
     }
 
     @Test
@@ -166,10 +166,10 @@ public class BankTests extends AbstractTest {
         driver.findElement(By.id("do_transaction")).click();
 
         // check new balance
-        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50.000,00 €");
+        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50,000.00 €");
         Assertions.assertEquals("-100", driver.findElement(By.id("betrag")).getAttribute("value"));
-        assertText(By.id("stat_money"), "100.000,00 €");
-        assertText(By.id("stat_bank"), "50.000,00 €");
+        assertText(By.id("stat_money"), "100,000.00 €");
+        assertText(By.id("stat_bank"), "50,000.00 €");
     }
 
     @Test
@@ -184,8 +184,8 @@ public class BankTests extends AbstractTest {
         assertElementPresent(By.id("meldung_110"));
 
         // check new balance
-        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50.000,00 €");
-        assertText(By.id("stat_money"), "100.000,00 €");
-        assertText(By.id("stat_bank"), "50.000,00 €");
+        assertText(By.id("cur_bank_account"), "Ihr Kontostand: 50,000.00 €");
+        assertText(By.id("stat_money"), "100,000.00 €");
+        assertText(By.id("stat_bank"), "50,000.00 €");
     }
 }