+++ tests/src/test/java/eu/fraho/blm2/st/PlantageTests.java
@@ -30,15 +30,15 @@ public class PlantageTests extends AbstractTest {
 
         driver.findElement(By.id("link_plantage")).click();
         setValue(By.id("stunden"), "2");
-        assertText(By.id("pr_ko_all"), "Kosten: 1.424,00 €");
+        assertText(By.id("pr_ko_all"), "Kosten: 1,424.00 €");
         driver.findElement(By.id("plant_all")).submit();
         assertElementPresent(By.id("meldung_207"));
         assertElementPresent(By.id("abort_1"));
         assertElementPresent(By.id("abort_2"));
-        assertText(By.id("stat_money"), "13.576,00 €");
+        assertText(By.id("stat_money"), "13,576.00 €");
 
         driver.findElement(By.id("link_buero")).click();
-        assertText(By.id("b_s_3"), "1.424,00 €");
+        assertText(By.id("b_s_3"), "1,424.00 €");
     }
 
     @Test
@@ -47,7 +47,7 @@ public class PlantageTests extends AbstractTest {
 
         driver.findElement(By.id("link_plantage")).click();
         setValue(By.id("stunden"), "13");
-        assertText(By.id("pr_ko_all"), "Kosten: 9.256,00 €");
+        assertText(By.id("pr_ko_all"), "Kosten: 9,256.00 €");
         driver.findElement(By.id("plant_all")).submit();
         assertElementPresent(By.id("meldung_133"));
     }
@@ -62,7 +62,7 @@ public class PlantageTests extends AbstractTest {
         assertElementPresent(By.id("meldung_207"));
         assertElementPresent(By.id("abort_1"));
         assertElementPresent(By.id("plant_2"));
-        assertText(By.id("stat_money"), "12.504,00 €");
+        assertText(By.id("stat_money"), "12,504.00 €");
     }
 
     @Test
@@ -75,7 +75,7 @@ public class PlantageTests extends AbstractTest {
         assertElementPresent(By.id("meldung_125"));
         assertElementPresent(By.id("plant_1"));
         assertElementPresent(By.id("plant_2"));
-        assertText(By.id("stat_money"), "15.000,00 €");
+        assertText(By.id("stat_money"), "15,000.00 €");
     }
 
     @Test
@@ -86,7 +86,7 @@ public class PlantageTests extends AbstractTest {
         setValue(By.id("amount_15"), "3");
         driver.findElement(By.id("plant_15")).submit();
         assertElementPresent(By.id("meldung_207"));
-        assertText(By.id("stat_money"), "14.999,22 €");
+        assertText(By.id("stat_money"), "14,999.22 €");
         assertElementPresent(By.id("plant_1"));
         assertElementPresent(By.id("abort_15"));
         Thread.sleep(TimeUnit.SECONDS.toMillis(3));
@@ -94,7 +94,7 @@ public class PlantageTests extends AbstractTest {
         driver.switchTo().alert().accept();
         assertElementPresent(By.id("meldung_222"));
         assertElementPresent(By.id("plant_15"));
-        assertText(By.id("stat_money"), "14.999,22 €");
+        assertText(By.id("stat_money"), "14,999.22 €");
 
         driver.findElement(By.id("link_bioladen")).click();
         assertText(By.id("cur_amount_15"), "1 kg");