+++ tests/src/test/java/eu/fraho/blm2/st/ResearchTests.java
@@ -38,20 +38,20 @@ public class ResearchTests extends AbstractTest {
 
         driver.findElement(By.id("link_forschungszentrum")).click();
 
-        assertText(By.id("stat_money"), "5.000,00 €");
+        assertText(By.id("stat_money"), "5,000.00 €");
         driver.findElement(By.id("research_1")).click();
         assertElementPresent(By.id("meldung_207"));
         assertElementPresent(By.id("abort_1"));
 
-        assertText(By.id("stat_money"), "4.517,26 €");
+        assertText(By.id("stat_money"), "4,517.26 €");
         driver.findElement(By.id("abort_1")).click();
         driver.switchTo().alert().accept();
         assertElementPresent(By.id("meldung_222"));
 
         assertElementPresent(By.id("research_1"));
-        assertText(By.id("stat_money"), "4.879,32 €");
+        assertText(By.id("stat_money"), "4,879.32 €");
 
         driver.findElement(By.id("link_buero")).click();
-        assertText(By.id("b_s_2"), "120,68 €");
+        assertText(By.id("b_s_2"), "120.68 €");
     }
 }