+++ tests/src/test/java/eu/fraho/blm2/st/ShopTests.java
@@ -28,14 +28,14 @@ public class ShopTests extends AbstractTest {
         WebDriver driver = getDriver();
 
         driver.findElement(By.id("link_bioladen")).click();
-        assertText(By.id("stat_money"), "5.000,00 €");
+        assertText(By.id("stat_money"), "5,000.00 €");
         driver.findElement(By.id("sell_all")).click();
 
         assertElementPresent(By.id("meldung_208"));
-        assertText(By.id("stat_money"), "5.374,64 €");
+        assertText(By.id("stat_money"), "5,374.64 €");
 
         driver.findElement(By.id("link_buero")).click();
-        assertText(By.id("b_i_1"), "374,64 €");
+        assertText(By.id("b_i_1"), "374.64 €");
     }
 
     @Test
@@ -43,18 +43,18 @@ public class ShopTests extends AbstractTest {
         WebDriver driver = getDriver();
 
         driver.findElement(By.id("link_bioladen")).click();
-        assertText(By.id("stat_money"), "5.000,00 €");
+        assertText(By.id("stat_money"), "5,000.00 €");
         setValue(By.id("amount_1"), "30");
         driver.findElement(By.id("sell_1")).click();
 
         assertElementPresent(By.id("meldung_208"));
-        assertText(By.id("stat_money"), "5.054,60 €");
+        assertText(By.id("stat_money"), "5,054.60 €");
         assertText(By.id("cur_amount_1"), "70 kg");
 
         driver.findElement(By.id("sell_1")).click();
-        assertText(By.id("stat_money"), "5.182,00 €");
+        assertText(By.id("stat_money"), "5,182.00 €");
 
         driver.findElement(By.id("link_buero")).click();
-        assertText(By.id("b_i_1"), "182,00 €");
+        assertText(By.id("b_i_1"), "182.00 €");
     }
 }